ai-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
spacy==3.7.2
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
pytesseract==0.3.10
Pillow==10.1.0
pandas==2.1.4
pydantic==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
numpy==1.26.2
scikit-learn==1.3.2
cryptography==41.0.7
python-magic==0.4.27
openai==1.3.5
anthropic==0.39.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
imaplib2==3.6
email-validator==2.1.0
exchangelib==5.1.0
boto3==1.34.10
